HR Function
HUMAN RESOURCE FUNCTIONS
The HR Function ensures the well-being and comfort of the employees in the working environment and it’s responsible for keeping track and managing of about seven different functions which are recruitment and selection, retention, training, planning, Human Resource data and analytics and compensation and benefits. All this functions have to be performed for the productivity of both the employees and the organization. To improve the efficiency of the HR Function the communication within a organization has to be improved to a standard where all employees are satisfied and also by enhancing the the training program. Provide feedback to employees continuously and by using as much technology as possible to ensure an effective human resource system.